requirements.txt
numpy==2.2.6
matplotlib==3.10.3
scipy==1.15.3
neuraloperator==1.0.2